added more content blocks. content blocks still need to be styled, only added basic coding for structure
Hey hey, so these are the notes for the code right now:

1. For Hero Content Block:

For the background image I usually put the class:
.background-image{
    background-size: cover;
    background-position: center;
    background-repeat: no-repeat;
}

This class with the min height of 90vh is on the background image at the moment. But this does not quite like the design. It’s a little off centre. So I could put a hero block class like this:

.hero-background{
	background-size: cover;
    	height: 90vh;
    	background-position: 10% 61%;
    	background-repeat: no-repeat;
}

The problem with this though is that this setting is specific to the image. How would you like to go about it considering they may change the image in the future.

2. For the footer

We need a logo with a paler palm tree as the palm tree blends with the background right now.